airflow/requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.7.2
async-timeout==4.0.2
autopep8==2.0.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.2.7
django-filter==21.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.4
environs==9.5.0
fastapi==0.86.0
gunicorn==20.0.4
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.20.1
packaging==21.3
prompt-toolkit==3.0.32
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pydantic==1.10.2
PyJWT==2.8.0
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2022.6
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xmltodict==0.13.0
